+++ blocks/cards/cards.js
@@ -1,10 +1,12 @@
 /* eslint-disable max-len */
-import { getRelativePath, getGenericIndexData } from '../../scripts/utils.js';
+import { getRelativePath, getGenericIndexData, generatePagination } from '../../scripts/utils.js';
 import { createOptimizedPicture } from '../../scripts/aem.js';
 
 const indexData = await getGenericIndexData();
+let imgWidth = '750';
 
-export function populateCard(container, cardInfo, imgWidth = '750') {
+/** function to populate card */
+export function populateCard(container, cardInfo) {
   const card = document.createElement('div');
   let efficacyBadge = '';
   if (cardInfo.efficacy) {
@@ -30,19 +32,55 @@ export function populateCard(container, cardInfo, imgWidth = '750') {
   container.append(card);
 }
 
-function renderCard(wrapper, link, imgWidth) {
-  const path = link ? link.getAttribute('href') : wrapper.textContent.trim();
-  const relPath = getRelativePath(path);
-  const cardInfo = indexData.find((item) => item.path === relPath);
-  if (cardInfo) {
-    const card = document.createElement('div');
-    populateCard(card, cardInfo, imgWidth);
-    wrapper.append(card);
+/** function to render card list when an array of card objects are passed.
+ * this also supports pagination
+*/
+export async function renderCardList(wrapper, cards, limit = 9) {
+  let limitPerPage = limit;
+  if (limit === undefined) {
+    limitPerPage = cards.length;
+  } else {
+    limitPerPage = Number.isNaN(parseInt(limit, 10)) ? 10 : parseInt(limit, 10);
+  }
+
+  wrapper.innerHTML = '';
+  let pageSize = 10;
+  if (!cards || cards.length === 0) {
+    return;
+  }
+  if (limitPerPage) {
+    pageSize = Math.round(limitPerPage - (limitPerPage % 3));
+  }
+  const list = document.createElement('div');
+  list.classList.add('article-teaser-list');
+  let currentPage = 1;
+  const match = window.location.hash.match(/page=(\d+)/);
+  if (match) {
+    currentPage = Number.isNaN(parseInt(match[1], 10)) ? currentPage : parseInt(match[1], 10);
+  }
+  const totalPages = Math.ceil(cards.length / pageSize);
+  const cardsList = cards.slice((currentPage - 1) * pageSize, currentPage * pageSize);
+
+  cardsList.forEach((card) => {
+    populateCard(wrapper, card);
+  });
+
+  if (totalPages > 1) {
+    const paginationContainer = document.createElement('div');
+    paginationContainer.classList.add('pagination-container');
+    paginationContainer.appendChild(generatePagination(currentPage, totalPages));
+    wrapper.append(paginationContainer);
   }
 }
 
+/** function to get an array of card objects from indexData */
+function getCardObject(link) {
+  const path = link?.getAttribute('href');
+  const relPath = getRelativePath(path);
+  return indexData.find((item) => item.path === relPath);
+}
+
 export default function decorate(block) {
-  let imgWidth = '750';
   if (block.classList.contains('suggested-videos')) {
     imgWidth = '200';
   }
@@ -51,7 +89,15 @@ export default function decorate(block) {
   const cardsWrapper = document.createElement('div');
   cardsWrapper.className = 'card-wrapper';
   block.innerHTML = '';
-  Array.from(cardLinks).map(async (cardLink) => renderCard(cardsWrapper, cardLink, imgWidth));
+  const cardsArray = [];
+  Array.from(cardLinks).map((cardLink) => {
+    const card = getCardObject(cardLink);
+    if (card) {
+      cardsArray.push(card);
+    }
+    return card;
+  });
+  renderCardList(cardsWrapper, cardsArray, undefined);
   if (title) {
     block.append(title);
   }
